Remove the persistence override.
In Ruby 3, this actually works as intended.
`prepend` was changed so that receivers of the
include also receive the `prepend`. This means
that this logic now executes on every AR create
and update. This is undesirable for us because
the method is inefficient right now (loops through
all attributes and we have defined a `translated?`
method on one of our models that directly conflicts
with persistence overrides.